printed out ea for move.b
in group_1_opcode_damian file, I used jsr and printed out the EAs. The
problems with it right now is that it is included in the overall jmp
table opcode. When you want to JSR into the jump table ea you have to
have the mode in d6, and the register and d7